logica 010623

+++ catalogo.js
@@ -1,5 +1,8 @@
 const cards = document.querySelector(".cards");
 
+var verificaemail;
+logarUsuario();
+
 carregarCatalogo();
 function carregarCatalogo(){
     let dados = JSON.parse(localStorage.getItem("catalogo"));
@@ -11,6 +14,7 @@ function carregarCatalogo(){
     }
 
     dados.forEach((elemento, indice) => {
+        if(elemento.email == verificaemail){
         let divcard = document.createElement("div");
         divcard.setAttribute("class", "card")
         divcard.innerHTML = `<div class="cardimg">
@@ -23,24 +27,33 @@ function carregarCatalogo(){
         <a onclick="editar(${indice})">editar</a>
         </div>`;
         
-        cards.appendChild(divcard);
+        cards.appendChild(divcard);}
         
     });
 }
 
 function excluir(indice){
+    if (confirm("Tem certeza que deseja excluir?")) {
     let dados = JSON.parse(localStorage.getItem("catalogo"));
-    if(dados.length == 1)
-    {localStorage.clear("catalogo");}
-    else{
     dados.splice(indice,1);
     localStorage.setItem("catalogo", JSON.stringify(dados));
-    }
     window.location.reload();
+    }
 }
 
 function editar(indice){
-    var url ="item.html?peditar=true&indice="+ encodeURIComponent(indice);
+    var url ="item.html?peditar=true&indice="+ 
+    encodeURIComponent(indice);
     window.location.href = url;
 }
 
+function logarUsuario(){
+    let email = sessionStorage.getItem("logado");
+      if (email != null) {
+        verificaemail = email;
+      } else {
+        window.location.assign("login.html");
+      }
+
+
+}
